+++ simpleadb/adbprocess.py
@@ -9,7 +9,7 @@
 """ Interface for adb process"""
 import subprocess
 from subprocess import CalledProcessError
-from typing import Optional
+from typing import List, Optional
 from . import adbcmds
 
 
@@ -38,35 +38,46 @@ class AdbCommandError(Exception):
 
 
 class AdbProcess:
-    """AdbProcess this class is used to call adb process
+    """ AdbProcess this class is used to call adb process.
 
-      Args:
-        path (Optional[str]): adb path, default: 'adb'
-      Returns:
-        0 if success, otherwise error code
+    :param Optional[str] device_id: Device ID, used when called adb command on
+        device.
+    :param: adb_path (Optional[str]): adb path, default: 'adb'
     """
 
-    def __init__(self, path: Optional[str] = adbcmds.ADB):
-        self.__adb_path = path
+    def __init__(self, device_id: Optional[str] = None,
+                 adb_path: Optional[str] = adbcmds.ADB):
+        self.device_id = device_id
+        self.adb_path = adb_path
 
-    def check_output(self, args: str, **kwargs) -> str:
-        """ Call process
+    def create_use_on_device_arg(self) -> str:
+        """ Create use on device argument.
 
-          Args:
-            args (str): Arguments
-            **kwargs: Arbitrary keyword arguments
-          Keyword Args:
-            timeout (int): Timeout in sec
-          Returns:
-            0 if success
-          Raises:
-            CalledProcessError: when failed
+        :return: Concatenated string used for adb commands '-s <device_id>', or
+            empty string.
+        :rtype: str
         """
-        cmd = ' '.join([
-            self.__adb_path,
-            args,
-        ])
+        return '-s ' + \
+            str(self.device_id) if self.device_id is not None else ''
+
+    def check_output(self, args: List[str], **kwargs) -> str:
+        """ Call adb subprocess.
+
+        :param List[str] prop: Arguments.
+        :keyword str timeout: Timeout in sec.
+        :raise: AdbCommandError: When failed.
+        :return: Process output.
+        """
+        cmd = []
+        cmd_args = [self.adb_path]
+        if self.device_id is not None:
+            cmd_args += [self.create_use_on_device_arg()]
+        cmd_args += args
+        cmd = ' '.join(cmd_args)
         kwargs.setdefault('shell', True)
         kwargs.setdefault('universal_newlines', True)
         kwargs.setdefault('stderr', subprocess.STDOUT)
-        return subprocess.check_output(cmd, **kwargs)
+        try:
+            return subprocess.check_output(cmd, **kwargs).rstrip('\n\r')
+        except CalledProcessError as err:
+            raise AdbCommandError(self.device_id, None, err) from err